requirements/base.txt
Django>=4.2
djangorestframework
psycopg2-binary
redis
celery
pytest
pytest-django
